requirements.txt
redis
Flask
requests
urllib3
dumper
waitress
pytest
pyyaml